Keep pool formulas near counterparty pools
