+++ src/utils/blockchain/counterparty/__tests__/pool.test.ts
@@ -1,0 +1,25 @@
+import { describe, expect, it } from "vitest";
+import {
+  applyPoolSlippage,
+  calculateInitialLpEstimate,
+  calculateLimitingLpEstimate,
+} from "../pool";
+
+describe("counterparty pool utilities", () => {
+  it("applies pool slippage to raw integer quantities", () => {
+    expect(applyPoolSlippage("100000000", "2.5")).toBe("97500000");
+    expect(applyPoolSlippage("100", "0.5")).toBe("99");
+    expect(applyPoolSlippage(undefined, "1")).toBe("0");
+  });
+
+  it("estimates initial LP supply from the geometric mean", () => {
+    expect(calculateInitialLpEstimate("100", "400")).toBe("200");
+    expect(calculateInitialLpEstimate("0", "400")).toBe("0");
+  });
+
+  it("uses the limiting side when a deposit is below the quoted ratio", () => {
+    expect(calculateLimitingLpEstimate("1000", "500", "250")).toBe("500");
+    expect(calculateLimitingLpEstimate("1000", "500", "500")).toBe("1000");
+    expect(calculateLimitingLpEstimate("1000", null, "250")).toBe("1000");
+  });
+});
